Fix using long instead of int in compute

+++ src/main/java/javaben/basic/Addition.java
@@ -5,7 +5,7 @@ public class Addition extends Basic {
     @Override
     public void compute(long n) {
         long a = 0;
-        for (int i = 0; i < n; i++) {
+        for (long i = 0; i < n; i++) {
             a += 7;
         }
     }
